+++ glue/src/main/java/org/somda/sdc/glue/consumer/report/ReportProcessor.java
@@ -19,6 +19,7 @@ import org.somda.sdc.glue.consumer.ConsumerConfig;
 import org.somda.sdc.glue.consumer.report.helper.ReportWriter;
 
 import javax.annotation.Nullable;
+import java.math.BigInteger;
 import java.net.URI;
 import java.util.concurrent.ArrayBlockingQueue;
 import java.util.concurrent.BlockingQueue;
@@ -145,7 +146,7 @@ public class ReportProcessor extends AbstractIdleService {
         final MdibVersion mdibAccessMdibVersion = mdibAccess.getMdibVersion();
         final MdibVersion reportMdibVersion = mdibVersionUtil.getMdibVersion(report);
         if (!mdibAccessMdibVersion.getSequenceId().equals(reportMdibVersion.getSequenceId()) ||
-                !mdibAccessMdibVersion.getInstanceId().equals(reportMdibVersion.getInstanceId())) {
+                !equalsInstanceIds(mdibAccessMdibVersion.getInstanceId(), reportMdibVersion.getInstanceId())) {
             throw new ReportProcessingException(String.format("MDIB version from MDIB (%s) and " +
                             "MDIB version from report (%s) do not match",
                     mdibAccessMdibVersion, reportMdibVersion));
@@ -207,7 +208,7 @@ public class ReportProcessor extends AbstractIdleService {
 
         final MdibVersion mdibVersion = mdibAccess.getMdibVersion();
         if (!URI.create(mdibVersion.getSequenceId()).equals(URI.create(contextStatesResponse.getSequenceId())) ||
-                !mdibVersion.getInstanceId().equals(contextStatesResponse.getInstanceId())) {
+                !equalsInstanceIds(mdibVersion.getInstanceId(), contextStatesResponse.getInstanceId())) {
             throw new ReportProcessingException(String.format("Received context state report belongs to " +
                             "different MDIB sequence/instance. Expected: %s, actual: %s",
                     mdibVersionUtil.getMdibVersion(contextStatesResponse), mdibVersion));
@@ -223,6 +224,20 @@ public class ReportProcessor extends AbstractIdleService {
         final MdibStateModifications modifications = MdibStateModifications.create(MdibStateModifications.Type.CONTEXT)
                 .addAll(contextStatesResponse.getContextState());
         mdibAccess.writeStates(mdibVersion, modifications);
+    }
+
+    /**
+     * Compares two InstanceIds for equality.
+     *
+     * This equals check is necessary, since InstanceId has a implied value of 0 and the report might actually have the
+     * field set to null.
+     *
+     * @param mdibVersionInstanceId the mdibVersion instance id
+     * @param reportInstanceId the report instance id, implied value always set to 0
+     * @return true if equal, false otherwise
+     */
+    private boolean equalsInstanceIds(BigInteger mdibVersionInstanceId, @Nullable BigInteger reportInstanceId) {
+        return reportInstanceId == null ? mdibVersionInstanceId.equals(BigInteger.ZERO) : mdibVersionInstanceId.equals(reportInstanceId);
     }
 
     @Override